Register, Login, Logout & Authorization for User (regarding own & others ideas)

+++ ideas-app/app/Http/Controllers/IdeaController.php
@@ -14,22 +14,24 @@ class IdeaController extends Controller
 
     public function store()
     {
-        request()->validate([
+
+        $validated = request()->validate([
             'content' => 'required|min:3|max:240',
 
         ]);
 
-        $idea = Idea::create(
-            [
-                'content' => request()->get('content', ''),
-            ]
-        );
+        $validated['user_id'] = auth()->id();
+
+        Idea::create($validated);
 
         return redirect()->route('dashboard')->with('success', 'Idea created successfully!');
     }
 
     public function destroy(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404);
+        }
         $idea->delete();
 
         return redirect()->route('dashboard')->with('success', 'Idea deleted successfully!');
@@ -37,6 +39,9 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+        if (auth()->id() !== $idea->user_id) {
+            abort(404);
+        }
         $editing = true;
 
         return view('ideas.show', compact('idea', 'editing'));
@@ -45,12 +50,14 @@ class IdeaController extends Controller
 
     public function update(Idea $idea)
     {
-        request()->validate([
+        if (auth()->id() !== $idea->user_id) {
+            abort(404);
+        }
+        $validated = request()->validate([
             'content' => 'required|min:3|max:240',
         ]);
 
-        $idea->content = request()->get('content', '');
-        $idea->save();
+        $idea->update($validated);
 
         return redirect()->route('ideas.show', $idea->id)->with('success', "Idea updated successfully!");
     }
